+++ src/Metadata/Operation.php
@@ -15,32 +15,50 @@ namespace ApiPlatform\Metadata;
 
 use ApiPlatform\State\OptionsInterface;
 
-abstract class Operation
+/**
+ * ⚠ This class and its children offer no backward compatibility regarding positional parameters.
+ */
+abstract class Operation extends Metadata
 {
     use WithResourceTrait;
-    /**
-     * @var mixed|null
-     */
-    protected mixed $input;
-    /**
-     * @var mixed|null
-     */
-    protected mixed $output;
-    /**
-     * @var string|callable|null
-     */
-    protected $provider;
-    /**
-     * @var string|callable|null
-     */
-    protected $processor;
 
     /**
-     * @param string[]   $filters
-     * @param mixed|null $input
-     * @param mixed|null $output
-     * @param mixed|null $provider
-     * @param mixed|null $processor
+     * @param bool|null                           $paginationEnabled              {@see https://api-platform.com/docs/core/pagination/#for-a-specific-resource}
+     * @param string|null                         $paginationType                 {@see https://api-platform.com/docs/core/graphql/#using-the-page-based-pagination}
+     * @param int|null                            $paginationItemsPerPage         {@see https://api-platform.com/docs/core/pagination/#changing-the-number-of-items-per-page}
+     * @param int|null                            $paginationMaximumItemsPerPage  {@see https://api-platform.com/docs/core/pagination/#changing-maximum-items-per-page}
+     * @param bool|null                           $paginationPartial              {@see https://api-platform.com/docs/core/performance/#partial-pagination}
+     * @param bool|null                           $paginationClientEnabled        {@see https://api-platform.com/docs/core/pagination/#for-a-specific-resource-1}
+     * @param bool|null                           $paginationClientItemsPerPage   {@see https://api-platform.com/docs/core/pagination/#for-a-specific-resource-3}
+     * @param bool|null                           $paginationClientPartial        {@see https://api-platform.com/docs/core/pagination/#for-a-specific-resource-6}
+     * @param bool|null                           $paginationFetchJoinCollection  {@see https://api-platform.com/docs/core/pagination/#controlling-the-behavior-of-the-doctrine-orm-paginator}
+     * @param array<string, string>|string[]|null $order                          {@see https://api-platform.com/docs/core/default-order/#overriding-default-order}
+     * @param string|null                         $security                       {@see https://api-platform.com/docs/core/security}
+     * @param string|null                         $securityMessage                {@see https://api-platform.com/docs/core/security/#configuring-the-access-control-error-message}
+     * @param string|null                         $securityPostDenormalize        {@see https://api-platform.com/docs/core/security/#executing-access-control-rules-after-denormalization}
+     * @param string|null                         $securityPostDenormalizeMessage {@see https://api-platform.com/docs/core/security/#configuring-the-access-control-error-message}
+     * @param string|null                         $deprecationReason              {@see https://api-platform.com/docs/core/deprecations/#deprecating-resource-classes-operations-and-properties}
+     * @param string[]|null                       $filters                        {@see https://api-platform.com/docs/core/filters/#doctrine-orm-and-mongodb-odm-filters}
+     * @param array{
+     *     class?: string|null,
+     *     name?: string,
+     * }|string|false|null $input {@see https://api-platform.com/docs/core/dto/#specifying-an-input-or-an-output-data-representation}
+     * @param array{
+     *     class?: string|null,
+     *     name?: string,
+     * }|string|false|null $output {@see https://api-platform.com/docs/core/dto/#specifying-an-input-or-an-output-data-representation}
+     * @param string|array|bool|null $mercure       {@see https://api-platform.com/docs/core/mercure}
+     * @param string|bool|null       $messenger     {@see https://api-platform.com/docs/core/messenger/#dispatching-a-resource-through-the-message-bus}
+     * @param bool|null              $elasticsearch {@see https://api-platform.com/docs/core/elasticsearch/}
+     * @param bool|null              $read          {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null              $deserialize   {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null              $validate      {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null              $write         {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null              $serialize     {@see https://api-platform.com/docs/core/events/#the-event-system}
+     * @param bool|null              $fetchPartial  {@see https://api-platform.com/docs/core/performance/#fetch-partial}
+     * @param bool|null              $forceEager    {@see https://api-platform.com/docs/core/performance/#force-eager}
+     * @param string|callable|null   $provider      {@see https://api-platform.com/docs/core/state-providers/#state-providers}
+     * @param string|callable|null   $processor     {@see https://api-platform.com/docs/core/state-processors/#state-processors}
      */
     public function __construct(
         protected ?string $shortName = null,
@@ -55,7 +73,6 @@ abstract class Operation
         protected ?bool $paginationClientPartial = null,
         protected ?bool $paginationFetchJoinCollection = null,
         protected ?bool $paginationUseOutputWalkers = null,
-        protected ?array $paginationViaCursor = null,
         protected ?array $order = null,
         protected ?string $description = null,
         protected ?array $normalizationContext = null,
@@ -70,15 +87,9 @@ abstract class Operation
         protected ?string $deprecationReason = null,
         protected ?array $filters = null,
         protected ?array $validationContext = null,
-        $input = null,
-        $output = null,
-        /**
-         * @var string|array|bool|null
-         */
+        protected $input = null,
+        protected $output = null,
         protected $mercure = null,
-        /**
-         * @var string|bool|null
-         */
         protected $messenger = null,
         protected ?bool $elasticsearch = null,
         protected ?int $urlGenerationStrategy = null,
@@ -91,465 +102,56 @@ abstract class Operation
         protected ?bool $forceEager = null,
         protected ?int $priority = null,
         protected ?string $name = null,
-        $provider = null,
-        $processor = null,
+        protected $provider = null,
+        protected $processor = null,
         protected ?OptionsInterface $stateOptions = null,
         protected array $extraProperties = [],
     ) {
-        $this->input = $input;
-        $this->output = $output;
-        $this->provider = $provider;
-        $this->processor = $processor;
+        parent::__construct(
+            shortName: $shortName,
+            class: $class,
+            description: $description,
+            urlGenerationStrategy: $urlGenerationStrategy,
+            deprecationReason: $deprecationReason,
+            normalizationContext: $normalizationContext,
+            denormalizationContext: $denormalizationContext,
+            collectDenormalizationErrors: $collectDenormalizationErrors,
+            validationContext: $validationContext,
+            filters: $filters,
+            elasticsearch: $elasticsearch,
+            mercure: $mercure,
+            messenger: $messenger,
+            input: $input,
+            output: $output,
+            order: $order,
+            fetchPartial: $fetchPartial,
+            forceEager: $forceEager,
+            paginationEnabled: $paginationEnabled,
+            paginationType: $paginationType,
+            paginationItemsPerPage: $paginationItemsPerPage,
+            paginationMaximumItemsPerPage: $paginationMaximumItemsPerPage,
+            paginationPartial: $paginationPartial,
+            paginationClientEnabled: $paginationClientEnabled,
+            paginationClientItemsPerPage: $paginationClientItemsPerPage,
+            paginationClientPartial: $paginationClientPartial,
+            paginationFetchJoinCollection: $paginationFetchJoinCollection,
+            paginationUseOutputWalkers: $paginationUseOutputWalkers,
+            security: $security,
+            securityMessage: $securityMessage,
+            securityPostDenormalize: $securityPostDenormalize,
+            securityPostDenormalizeMessage: $securityPostDenormalizeMessage,
+            securityPostValidation: $securityPostValidation,
+            securityPostValidationMessage: $securityPostValidationMessage,
+            provider: $provider,
+            processor: $processor,
+            stateOptions: $stateOptions,
+            extraProperties: $extraProperties,
+        );
     }
 
     public function withOperation($operation)
     {
         return $this->copyFrom($operation);
-    }
-
-    public function getShortName(): ?string
-    {
-        return $this->shortName;
-    }
-
-    public function withShortName(?string $shortName = null): self
-    {
-        $self = clone $this;
-        $self->shortName = $shortName;
-
-        return $self;
-    }
-
-    public function getClass(): ?string
-    {
-        return $this->class;
-    }
-
-    public function withClass(?string $class = null): self
-    {
-        $self = clone $this;
-        $self->class = $class;
-
-        return $self;
-    }
-
-    public function getPaginationEnabled(): ?bool
-    {
-        return $this->paginationEnabled;
-    }
-
-    public function withPaginationEnabled(?bool $paginationEnabled = null): self
-    {
-        $self = clone $this;
-        $self->paginationEnabled = $paginationEnabled;
-
-        return $self;
-    }
-
-    public function getPaginationType(): ?string
-    {
-        return $this->paginationType;
-    }
-
-    public function withPaginationType(?string $paginationType = null): self
-    {
-        $self = clone $this;
-        $self->paginationType = $paginationType;
-
-        return $self;
-    }
-
-    public function getPaginationItemsPerPage(): ?int
-    {
-        return $this->paginationItemsPerPage;
-    }
-
-    public function withPaginationItemsPerPage(?int $paginationItemsPerPage = null): self
-    {
-        $self = clone $this;
-        $self->paginationItemsPerPage = $paginationItemsPerPage;
-
-        return $self;
-    }
-
-    public function getPaginationMaximumItemsPerPage(): ?int
-    {
-        return $this->paginationMaximumItemsPerPage;
-    }
-
-    public function withPaginationMaximumItemsPerPage(?int $paginationMaximumItemsPerPage = null): self
-    {
-        $self = clone $this;
-        $self->paginationMaximumItemsPerPage = $paginationMaximumItemsPerPage;
-
-        return $self;
-    }
-
-    public function getPaginationPartial(): ?bool
-    {
-        return $this->paginationPartial;
-    }
-
-    public function withPaginationPartial(?bool $paginationPartial = null): self
-    {
-        $self = clone $this;
-        $self->paginationPartial = $paginationPartial;
-
-        return $self;
-    }
-
-    public function getPaginationClientEnabled(): ?bool
-    {
-        return $this->paginationClientEnabled;
-    }
-
-    public function withPaginationClientEnabled(?bool $paginationClientEnabled = null): self
-    {
-        $self = clone $this;
-        $self->paginationClientEnabled = $paginationClientEnabled;
-
-        return $self;
-    }
-
-    public function getPaginationClientItemsPerPage(): ?bool
-    {
-        return $this->paginationClientItemsPerPage;
-    }
-
-    public function withPaginationClientItemsPerPage(?bool $paginationClientItemsPerPage = null): self
-    {
-        $self = clone $this;
-        $self->paginationClientItemsPerPage = $paginationClientItemsPerPage;
-
-        return $self;
-    }
-
-    public function getPaginationClientPartial(): ?bool
-    {
-        return $this->paginationClientPartial;
-    }
-
-    public function withPaginationClientPartial(?bool $paginationClientPartial = null): self
-    {
-        $self = clone $this;
-        $self->paginationClientPartial = $paginationClientPartial;
-
-        return $self;
-    }
-
-    public function getPaginationFetchJoinCollection(): ?bool
-    {
-        return $this->paginationFetchJoinCollection;
-    }
-
-    public function withPaginationFetchJoinCollection(?bool $paginationFetchJoinCollection = null): self
-    {
-        $self = clone $this;
-        $self->paginationFetchJoinCollection = $paginationFetchJoinCollection;
-
-        return $self;
-    }
-
-    public function getPaginationUseOutputWalkers(): ?bool
-    {
-        return $this->paginationUseOutputWalkers;
-    }
-
-    public function withPaginationUseOutputWalkers(?bool $paginationUseOutputWalkers = null): self
-    {
-        $self = clone $this;
-        $self->paginationUseOutputWalkers = $paginationUseOutputWalkers;
-
-        return $self;
-    }
-
-    public function getPaginationViaCursor(): ?array
-    {
-        return $this->paginationViaCursor;
-    }
-
-    public function withPaginationViaCursor(array $paginationViaCursor): self
-    {
-        $self = clone $this;
-        $self->paginationViaCursor = $paginationViaCursor;
-
-        return $self;
-    }
-
-    public function getOrder(): ?array
-    {
-        return $this->order;
-    }
-
-    public function withOrder(array $order = []): self
-    {
-        $self = clone $this;
-        $self->order = $order;
-
-        return $self;
-    }
-
-    public function getDescription(): ?string
-    {
-        return $this->description;
-    }
-
-    public function withDescription(?string $description = null): self
-    {
-        $self = clone $this;
-        $self->description = $description;
-
-        return $self;
-    }
-
-    public function getNormalizationContext(): ?array
-    {
-        return $this->normalizationContext;
-    }
-
-    public function withNormalizationContext(array $normalizationContext = []): self
-    {
-        $self = clone $this;
-        $self->normalizationContext = $normalizationContext;
-
-        return $self;
-    }
-
-    public function getDenormalizationContext(): ?array
-    {
-        return $this->denormalizationContext;
-    }
-
-    public function withDenormalizationContext(array $denormalizationContext = []): self
-    {
-        $self = clone $this;
-        $self->denormalizationContext = $denormalizationContext;
-
-        return $self;
-    }
-
-    public function getCollectDenormalizationErrors(): ?bool
-    {
-        return $this->collectDenormalizationErrors;
-    }
-
-    public function withCollectDenormalizationErrors(bool $collectDenormalizationErrors = null): self
-    {
-        $self = clone $this;
-        $self->collectDenormalizationErrors = $collectDenormalizationErrors;
-
-        return $self;
-    }
-
-    public function getSecurity(): ?string
-    {
-        return $this->security;
-    }
-
-    public function withSecurity(?string $security = null): self
-    {
-        $self = clone $this;
-        $self->security = $security;
-
-        return $self;
-    }
-
-    public function getSecurityMessage(): ?string
-    {
-        return $this->securityMessage;
-    }
-
-    public function withSecurityMessage(?string $securityMessage = null): self
-    {
-        $self = clone $this;
-        $self->securityMessage = $securityMessage;
-
-        return $self;
-    }
-
-    public function getSecurityPostDenormalize(): ?string
-    {
-        return $this->securityPostDenormalize;
-    }
-
-    public function withSecurityPostDenormalize(?string $securityPostDenormalize = null): self
-    {
-        $self = clone $this;
-        $self->securityPostDenormalize = $securityPostDenormalize;
-
-        return $self;
-    }
-
-    public function getSecurityPostDenormalizeMessage(): ?string
-    {
-        return $this->securityPostDenormalizeMessage;
-    }
-
-    public function withSecurityPostDenormalizeMessage(?string $securityPostDenormalizeMessage = null): self
-    {
-        $self = clone $this;
-        $self->securityPostDenormalizeMessage = $securityPostDenormalizeMessage;
-
-        return $self;
-    }
-
-    public function getSecurityPostValidation(): ?string
-    {
-        return $this->securityPostValidation;
-    }
-
-    public function withSecurityPostValidation(?string $securityPostValidation = null): self
-    {
-        $self = clone $this;
-        $self->securityPostValidation = $securityPostValidation;
-
-        return $self;
-    }
-
-    public function getSecurityPostValidationMessage(): ?string
-    {
-        return $this->securityPostValidationMessage;
-    }
-
-    public function withSecurityPostValidationMessage(?string $securityPostValidationMessage = null): self
-    {
-        $self = clone $this;
-        $self->securityPostValidationMessage = $securityPostValidationMessage;
-
-        return $self;
-    }
-
-    public function getDeprecationReason(): ?string
-    {
-        return $this->deprecationReason;
-    }
-
-    public function withDeprecationReason(?string $deprecationReason = null): self
-    {
-        $self = clone $this;
-        $self->deprecationReason = $deprecationReason;
-
-        return $self;
-    }
-
-    public function getFilters(): ?array
-    {
-        return $this->filters;
-    }
-
-    public function withFilters(array $filters = []): self
-    {
-        $self = clone $this;
-        $self->filters = $filters;
-
-        return $self;
-    }
-
-    public function getValidationContext(): ?array
-    {
-        return $this->validationContext;
-    }
-
-    public function withValidationContext(array $validationContext = []): self
-    {
-        $self = clone $this;
-        $self->validationContext = $validationContext;
-
-        return $self;
-    }
-
-    public function getInput()
-    {
-        return $this->input;
-    }
-
-    public function withInput($input = null): self
-    {
-        $self = clone $this;
-        $self->input = $input;
-
-        return $self;
-    }
-
-    public function getOutput()
-    {
-        return $this->output;
-    }
-
-    public function withOutput($output = null): self
-    {
-        $self = clone $this;
-        $self->output = $output;
-
-        return $self;
-    }
-
-    /**
-     * @return bool|string|array|null
-     */
-    public function getMercure()
-    {
-        return $this->mercure;
-    }
-
-    /**
-     * @param bool|string|array|null $mercure
-     *
-     * @return $this
-     */
-    public function withMercure($mercure = null): self
-    {
-        $self = clone $this;
-        $self->mercure = $mercure;
-
-        return $self;
-    }
-
-    /**
-     * @return bool|string|null
-     */
-    public function getMessenger()
-    {
-        return $this->messenger;
-    }
-
-    /**
-     * @param bool|string|null $messenger
-     *
-     * @return $this
-     */
-    public function withMessenger($messenger = null): self
-    {
-        $self = clone $this;
-        $self->messenger = $messenger;
-
-        return $self;
-    }
-
-    public function getElasticsearch(): ?bool
-    {
-        return $this->elasticsearch;
-    }
-
-    public function withElasticsearch(?bool $elasticsearch = null): self
-    {
-        $self = clone $this;
-        $self->elasticsearch = $elasticsearch;
-
-        return $self;
-    }
-
-    public function getUrlGenerationStrategy(): ?int
-    {
-        return $this->urlGenerationStrategy;
-    }
-
-    public function withUrlGenerationStrategy(?int $urlGenerationStrategy = null): self
-    {
-        $self = clone $this;
-        $self->urlGenerationStrategy = $urlGenerationStrategy;
-
-        return $self;
     }
 
     public function canRead(): ?bool
@@ -617,32 +219,6 @@ abstract class Operation
         return $self;
     }
 
-    public function getFetchPartial(): ?bool
-    {
-        return $this->fetchPartial;
-    }
-
-    public function withFetchPartial(?bool $fetchPartial = null): self
-    {
-        $self = clone $this;
-        $self->fetchPartial = $fetchPartial;
-
-        return $self;
-    }
-
-    public function getForceEager(): ?bool
-    {
-        return $this->forceEager;
-    }
-
-    public function withForceEager(?bool $forceEager = null): self
-    {
-        $self = clone $this;
-        $self->forceEager = $forceEager;
-
-        return $self;
-    }
-
     public function getPriority(): ?int
     {
         return $this->priority;
@@ -665,58 +241,6 @@ abstract class Operation
     {
         $self = clone $this;
         $self->name = $name;
-
-        return $self;
-    }
-
-    public function getProcessor(): callable|string|null
-    {
-        return $this->processor;
-    }
-
-    public function withProcessor(callable|string|null $processor): self
-    {
-        $self = clone $this;
-        $self->processor = $processor;
-
-        return $self;
-    }
-
-    public function getProvider(): callable|string|null
-    {
-        return $this->provider;
-    }
-
-    public function withProvider(callable|string|null $provider): self
-    {
-        $self = clone $this;
-        $self->provider = $provider;
-
-        return $self;
-    }
-
-    public function getExtraProperties(): array
-    {
-        return $this->extraProperties;
-    }
-
-    public function withExtraProperties(array $extraProperties = []): self
-    {
-        $self = clone $this;
-        $self->extraProperties = $extraProperties;
-
-        return $self;
-    }
-
-    public function getStateOptions(): ?OptionsInterface
-    {
-        return $this->stateOptions;
-    }
-
-    public function withStateOptions(?OptionsInterface $stateOptions): self
-    {
-        $self = clone $this;
-        $self->stateOptions = $stateOptions;
 
         return $self;
     }