+++ tests/Behat/DoctrineContext.php
@@ -65,6 +65,7 @@ use ApiPlatform\Tests\Fixtures\TestBundle\Document\IriOnlyDummy as IriOnlyDummyD
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\MaxDepthDummy as MaxDepthDummyDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\MultiRelationsDummy as MultiRelationsDummyDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\MultiRelationsRelatedDummy as MultiRelationsRelatedDummyDocument;
+use ApiPlatform\Tests\Fixtures\TestBundle\Document\MusicGroup as MusicGroupDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\NetworkPathDummy as NetworkPathDummyDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\NetworkPathRelationDummy as NetworkPathRelationDummyDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\Order as OrderDocument;
@@ -89,6 +90,7 @@ use ApiPlatform\Tests\Fixtures\TestBundle\Document\Taxon as TaxonDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\ThirdLevel as ThirdLevelDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\UrlEncodedId as UrlEncodedIdDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\User as UserDocument;
+use ApiPlatform\Tests\Fixtures\TestBundle\Document\VideoGame as VideoGameDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Document\WithJsonDummy as WithJsonDummyDocument;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\AbsoluteUrlDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\AbsoluteUrlRelationDummy;
@@ -143,6 +145,7 @@ use ApiPlatform\Tests\Fixtures\TestBundle\Entity\IriOnlyDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\MaxDepthDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\MultiRelationsDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\MultiRelationsRelatedDummy;
+use ApiPlatform\Tests\Fixtures\TestBundle\Entity\MusicGroup;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\NetworkPathDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\NetworkPathRelationDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\Order;
@@ -163,6 +166,7 @@ use ApiPlatform\Tests\Fixtures\TestBundle\Entity\RelatedOwningDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\RelatedSecuredDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\RelatedToDummyFriend;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\RelationEmbedder;
+use ApiPlatform\Tests\Fixtures\TestBundle\Entity\RelationMultiple;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\SecuredDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\SeparatedEntity;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\Site;
@@ -170,9 +174,11 @@ use ApiPlatform\Tests\Fixtures\TestBundle\Entity\SoMany;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\SymfonyUuidDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\Taxon;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\ThirdLevel;
+use ApiPlatform\Tests\Fixtures\TestBundle\Entity\TreeDummy;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\UrlEncodedId;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\User;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\UuidIdentifierDummy;
+use ApiPlatform\Tests\Fixtures\TestBundle\Entity\VideoGame;
 use ApiPlatform\Tests\Fixtures\TestBundle\Entity\WithJsonDummy;
 use Behat\Behat\Context\Context;
 use Behat\Gherkin\Node\PyStringNode;
@@ -799,6 +805,22 @@ final class DoctrineContext implements Context
             $this->manager->persist($relatedDummy);
             $this->manager->persist($dummy);
         }
+        $this->manager->flush();
+    }
+
+    /**
+     * @Given there are tree dummies
+     */
+    public function thereAreTreeDummies(): void
+    {
+        $parentDummy = new TreeDummy();
+        $this->manager->persist($parentDummy);
+
+        $childDummy = new TreeDummy();
+        $childDummy->setParent($parentDummy);
+
+        $this->manager->persist($childDummy);
+
         $this->manager->flush();
     }
 
@@ -2033,6 +2055,80 @@ final class DoctrineContext implements Context
             $entity->value = (string) $i;
             $this->manager->persist($entity);
         }
+        $this->manager->flush();
+    }
+
+    /**
+     * @Given there is a video game with music groups
+     */
+    public function thereAreVideoGamesWithMusicGroups(): void
+    {
+        $sum41 = $this->buildMusicGroup();
+        $sum41->name = 'Sum 41';
+        $this->manager->persist($sum41);
+        $franz = $this->buildMusicGroup();
+        $franz->name = 'Franz Ferdinand';
+        $this->manager->persist($franz);
+
+        $videoGame = $this->buildVideoGame();
+        $videoGame->name = 'Guitar Hero';
+        $videoGame->addMusicGroup($sum41);
+        $videoGame->addMusicGroup($franz);
+        $this->manager->persist($videoGame);
+        $this->manager->flush();
+    }
+
+    /**
+     * @Given there is a relationMultiple object
+     */
+    public function thereIsARelationMultipleObject(): void
+    {
+        $first = $this->buildDummy();
+        $first->setId(1);
+        $first->setName('foo');
+        $second = $this->buildDummy();
+        $second->setId(2);
+        $second->setName('bar');
+
+        $relationMultiple = (new RelationMultiple());
+        $relationMultiple->first = $first;
+        $relationMultiple->second = $second;
+
+        $this->manager->persist($first);
+        $this->manager->persist($second);
+        $this->manager->persist($relationMultiple);
+
+        $this->manager->flush();
+    }
+
+    /**
+     * @Given there is a dummy object with many multiple relation
+     */
+    public function thereIsADummyObjectWithManyMultipleRelation(): void
+    {
+        $first = $this->buildDummy();
+        $first->setId(1);
+        $first->setName('foo');
+        $second = $this->buildDummy();
+        $second->setId(2);
+        $second->setName('bar');
+        $third = $this->buildDummy();
+        $third->setId(3);
+        $third->setName('foobar');
+
+        $relationMultiple1 = (new RelationMultiple());
+        $relationMultiple1->first = $first;
+        $relationMultiple1->second = $second;
+
+        $relationMultiple2 = (new RelationMultiple());
+        $relationMultiple2->first = $first;
+        $relationMultiple2->second = $third;
+
+        $this->manager->persist($first);
+        $this->manager->persist($second);
+        $this->manager->persist($third);
+        $this->manager->persist($relationMultiple1);
+        $this->manager->persist($relationMultiple2);
 
         $this->manager->flush();
     }
@@ -2375,5 +2471,15 @@ final class DoctrineContext implements Context
     private function buildMultiRelationsRelatedDummy(): MultiRelationsRelatedDummy|MultiRelationsRelatedDummyDocument
     {
         return $this->isOrm() ? new MultiRelationsRelatedDummy() : new MultiRelationsRelatedDummyDocument();
+    }
+
+    private function buildMusicGroup(): MusicGroup|MusicGroupDocument
+    {
+        return $this->isOrm() ? new MusicGroup() : new MusicGroupDocument();
+    }
+
+    private function buildVideoGame(): VideoGame|VideoGameDocument
+    {
+        return $this->isOrm() ? new VideoGame() : new VideoGameDocument();
     }
 }